Change to Experiment Constructor
simulator_config parameter is now optional. If none is supplied,
the system will use the DEFAULT_SIMULATION config. Again, this
was done to avoid using the simulator config file in the
journal paper, and make them more simple.

ATTENTION: I think there is a severe bug in the current version of
the simulator; the DEFAULT_SIMULATION config uses attributes in all caps
(LIKE THIS) while the essentials.config file is small caps (like this).
If such file is loaded, it won't be used at all since the simulator
uses the attributes in all-caps. I fixed the instances in the experimentation
and plot_factory that were small-caps, but the configuration file needs
to be updated.