Allocator efficiency improvement
The sample allocators now check which nodes in the system are full
before each allocation. A node is considered full if there are no
available CPU or Memory resources, which are necessary for all jobs
to run. Full nodes are discarded from the node list, and not considered
during the allocation process. This has great impact on performance.

The change affects the FFP, BFP and Balanced allocators. Weighted and
Hybrid do not need this as they already implement similar mechanisms.