Minor fixes
- The from_isodatetime_2_timestamp method now consider the timezone to
adjust the time transformation. By default is UTC+0
- Experiment class fixed
- If a Dispatcher with a predefined allocator needs to skip jobs which
are not possible to allocate to the system at the current time, and need
to continue trying the remaining with the queued jobs the
"skip_jobs_on_allocation=True" parameter must be passed to the
SchedulerBase object. This works only when a dispatcher is created as
Scheduler(Allocator()) and it's automatically called.